+++ example/lib/src/list_app.dart
@@ -54,13 +54,13 @@ class PostsViewStateBuilder extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ViewStateBuilder<List<Post>, PostsBloc>(
-      onLoading: (context) => const LoadingIndicator(),
-      onSuccess: (context, posts) =>
+      loading: (context) => const LoadingIndicator(),
+      data: (context, posts) =>
           PostsList(posts, onRefresh: () => _refreshPosts(context)),
-      onRefreshing: (context, posts) =>
+      refreshing: (context, posts) =>
           PostsList(posts, onRefresh: () => _refreshPosts(context)),
-      onEmpty: (context) => const PostsListEmpty(),
-      onError: (context, error) => ErrorMessage(error: error),
+      empty: (context) => const PostsListEmpty(),
+      error: (context, error) => ErrorMessage(error: error),
     );
   }