+++ lib/src/tools/connection/connection_builder.dart
@@ -9,21 +9,28 @@ import 'package:flutter_bloc_patterns/src/tools/connection/connection_bloc.dart'
 /// It's a wrapper over the [BlocBuilder] widget so it accepts a [bloc] object
 /// and provides [WidgetBuilder] functions for possible states:
 ///
-/// [onOnline] - a builder for the the [Connection.online] state,
-/// [onOffline] - a builder for the [Connection.offline] state.
+/// [online] - a builder for the the [Connection.online] state,
+/// [offline] - a builder for the [Connection.offline] state.
 class ConnectionBuilder extends BlocBuilder<ConnectionBloc, Connection> {
   ConnectionBuilder({
     super.key,
     super.bloc,
+    @Deprecated('This builder will be removed. Use "online" instead.')
     WidgetBuilder? onOnline,
+    WidgetBuilder? online,
+    @Deprecated('This builder will be removed. Use "offline" instead.')
     WidgetBuilder? onOffline,
+    WidgetBuilder? offline,
   }) : super(
           builder: (BuildContext context, Connection state) {
+            const none = SizedBox.shrink();
             switch (state) {
               case Connection.online:
-                return onOnline?.call(context) ?? const SizedBox.shrink();
+                return (online?.call(context) ?? onOnline?.call(context)) ??
+                    none;
               case Connection.offline:
-                return onOffline?.call(context) ?? const SizedBox.shrink();
+                return (offline?.call(context) ?? onOffline?.call(context)) ??
+                    none;
             }
           },
         );